Moved different tests to different files
