+++ services/pseudo_controller/src/main/resources/application.properties
@@ -12,64 +12,10 @@
 #     FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
 #     WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-server.port=9014
+### Whether to test NCM, or to test Alcor HTTP APIs ###
+test_against_ncm = false
 
-# Node Manager Microservice url configuration
-microservices.node.service.url=http://localhost:9007/nodes
-microservices.dataplane.nodeservice.url=http://localhost:9010/nodes
-
-dataplane.grpc.port = 50001
-dataplane.isovs=true
-
-grpc.min-threads = 100
-grpc.max-threads = 200
-grpc.threads-pool-name = grpc-thread-pool
-grpc.number-of-channels-per-host = 10
-grpc.number-of-warmups-per-channel = 1
-grpc.monitor-hosts = 0.0.0.0,0.0.0.1
-
-#DPM v1 is statically using version 101
-#DPM v2 will start with version 102
-protobuf.goal-state-message.version = 102
-
-#####Ignite configuration######
-ignite.host=localhost
-ignite.port=10800
-ignite.thin.client.enable=true
-#ignite.key-store-path=keystore.jks
-#ignite.key-store-password=123456
-#ignite.trust-store-path=truststore.jks
-#ignite.trust-store-password=12345
-
-#####Spring health#####
-management.health.redis.enabled=false
-
-#####Logging configuration#####
-#logging.file.path=./
-#logging.file.name=port-manager.log
-logging.level.root=INFO
-
-##### Jaeger ####
-opentracing.jaeger.udp-sender.host = localhost
-opentracing.jaeger.udp-sender.port = 6831
-opentracing.jaeger.const-sampler.decision = true
-opentracing.jaeger.enabled = true
-opentracing.jaeger.log-spans = true
-opentracing.jaeger.enable-w3c-propagation = true
-opentracing.jaeger.enable-b3-propagation = true
-opentracing.jaeger.service-name = alcor-ncm
-
-#### Jaeger Span Names ####
-jaeger.span-name.ncm_server_pushdown_gs=alcor-ncm-server-pushdown-gs
-jaeger.span-name.ncm_server_store_gs=alcor-ncm-server-store-gs
-jaeger.span-name.ncm_server_filter_send_gs=alcor-ncm-server-filter-send-gs
-jaeger.span-name.ncm_server_reply_dpm=alcor-ncm-server-reply-dpm
-jaeger.span-name.ncm_server_on_demand=alcor-ncm-on-demand
-jaeger.span-name.ncm_server_on_demand_retrieve_gs=alcor-ncm-on-demand-retrieve-gs
-jaeger.span-name.ncm_server_on_demand_send_gs=alcor-ncm-on-demand-send-gs
-jaeger.span-name.ncm_server_on_demand_reply=alcor-ncm-on-demand-reply
-
-### Pseudo Controller Params ###
+### Test Controller NCM test Params ###
 number_of_vpcs = 10
 ports_node_one = 100
 ports_node_two = 200
@@ -81,9 +27,12 @@ user_name = ubuntu
 password = gfedcba
 ### 0 is concurrent, others are sequential ###
 ping_mode = 1
+### 1 to do background ping, otherwise NOT to do background ping ###
 background_ping = 1
+### 0 to create container and ping, otherwise NOT to create container and NOT to ping ###
 create_container_and_ping = 1
-test_against_ncm = false
+
+### Test Controller Alcor HTTP APIs Test Params ###
 vpm_ip = 127.0.0.1
 vpm_port = 8080
 snm_ip = 127.0.0.1
